dmaster copy/pyproject.toml
[tool.poetry]
name = "dmaster"
version = "0.1.0"
description = "Disease Mastery Ontology System"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.0"
numpy = "^1.21.6"
owlready2 = "^0.37"
rdflib = "^7.1.4"
networkx = "^3.1"
pyshacl = "^0.30.1"
python-dotenv = "^1.1.0"
kazu = "^2.3.0"
scipy = "^1.11.4"
thinc = "^8.2.5"
protobuf = "^6.31.1"
tiktoken = "^0.9.0"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
